remove keycloak role-based auth, use request_loader to get user from token
